pyproject.toml
[project]
name = "project2-newsference"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "asyncio>=3.4.3",
    "fastapi>=0.115.8",
    "jinja2>=3.1.5",
    "lxml-html-clean>=0.4.1",
    "newsapi-python>=0.2.7",
    "newspaper3k>=0.2.8",
    "openai>=1.61.0",
    "pandas>=2.2.3",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
    "uvicorn>=0.34.0",
]
